Support non-sequential files
Closes #1

Even for sequential files this should have been the approach, it's much cleaner.

We create a Vector, each element of the vector is one of:

* >5: This represents the _next_ page of the file
  (can't be <= 5 because of reserved pages)
* 1: This represents that the _current index_ is the last page of the file
* 3: This represents that the _current index_ is a free page

It's a linked list with '1' as nil, all we need to know is where to start.

Luckily the NoteTable has the startPage for each file!