[WIP]: Did some refactoring to closer match spec
I think the next step is to look at the actual offset (compressed/uncompressed)
and see if they make sense.

What I _believe_ is happening is that the offsets are from the beginning of the
_file/note_ and not the beginnig of the compression data.

If this is the case then it's an easy fix. If it's not the case, rip.